+++ lib/mpl_toolkits/tests/test_mplot3d.py
@@ -343,8 +343,8 @@ def test_poly3dcollection_closed():
     fig = plt.figure()
     ax = fig.gca(projection='3d')
 
-    poly1 = np.array([[0, 0, 1], [0, 1, 1], [0, 0, 0]], np.float)
-    poly2 = np.array([[0, 1, 1], [1, 1, 1], [1, 1, 0]], np.float)
+    poly1 = np.array([[0, 0, 1], [0, 1, 1], [0, 0, 0]], float)
+    poly2 = np.array([[0, 1, 1], [1, 1, 1], [1, 1, 0]], float)
     c1 = art3d.Poly3DCollection([poly1], linewidths=3, edgecolor='k',
                                 facecolor=(0.5, 0.5, 1, 0.5), closed=True)
     c2 = art3d.Poly3DCollection([poly2], linewidths=3, edgecolor='k',
